+++ src/main/java/cz/muni/fi/pv168/seminar01/beta/model/IDGenerator.java
@@ -4,19 +4,19 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class IDGenerator {
-    private static final Map<Integer, Integer> IDs = new HashMap<>();
+    private static final Map<Long, Long> IDs = new HashMap<>();
 
     /**
      * Creates unique ID for supported classes. Only Ride, Category, Passenger and Vehicle classes
      * are supported (this ensures uniqueness).
      * <p>
-     * It uses first letter (as int) from class name as a prefix.
+     * It uses first letter (as Long) from class name as a prefix.
      *
      * @param caller - One of the supported classes.
      * @return unique ID
      */
-    public static int getNewID(Class<?> caller) {
-        int prefix;
+    public static long getNewID(Class<?> caller) {
+        long prefix;
         if (
                 caller.equals(Ride.class) ||
                         caller.equals(RideCategory.class) ||
@@ -31,10 +31,10 @@ public class IDGenerator {
 
         if (IDs.containsKey(prefix)) {
             IDs.put(prefix, IDs.get(prefix) + 1);
-            return Integer.parseInt("" + prefix + IDs.get(prefix));
+            return Long.parseLong("" + prefix + IDs.get(prefix));
         } else {
-            IDs.put(prefix, 0);
-            return Integer.parseInt("" + prefix + 0);
+            IDs.put(prefix, 0L);
+            return Long.parseLong("" + prefix + 0);
         }
     }
 }
